@@ -1,3 +1,3 @@
 """Transforms draft-3 CWL documents into v1.0+ as idiomatically as possible."""
 
-__version__ = "1.2.12"
+__version__ = "1.2.13"
